build(exports): only class export allowed in component file
only class export allowed in component file. This required some Type name changes.

COMUI-286
