+++ src/components/beta/gux-breadcrumbs/gux-breadcrumbs.types.ts
@@ -1,0 +1,1 @@
+export type GuxBreadcrumbAccent = 'primary' | 'secondary';
