+++ src/components/beta/gux-radial-loading/gux-radial-loading.types.ts
@@ -1,0 +1,1 @@
+export type GuxRadialLoadingContext = 'full-page' | 'input' | 'modal';
