+++ src/components/stable/gux-button/gux-button.types.ts
@@ -1,0 +1,1 @@
+export type GuxButtonAccent = 'primary' | 'secondary';
