--- src/components/stable/gux-calendar/gux-calendar-constants.ts
@@ -1,6 +1,0 @@
-export interface IDateElement {
-  class: string;
-  date: Date;
-  selected: boolean;
-  hidden: boolean;
-}
