+++ src/components/stable/gux-datepicker/gux-datepicker.types.ts
@@ -1,0 +1,3 @@
+import { CalendarModes } from '../../../common-enums';
+
+export type GuxDatepickerMode = CalendarModes.Single | CalendarModes.Range;
