+++ src/components/stable/gux-disclosure-button/gux-disclosure-button.types.ts
@@ -1,0 +1,1 @@
+export type GuxDisclosureButtonPosition = 'left' | 'right';
