+++ src/components/stable/gux-modal/gux-modal.types.ts
@@ -1,0 +1,1 @@
+export type GuxModalSize = 'small' | 'medium' | 'large';
