+++ src/components/stable/gux-notification-toast/gux-notification-toast.types.ts
@@ -1,0 +1,5 @@
+export type GuxNotificationToastAccent =
+  | 'neutral'
+  | 'positive'
+  | 'alert'
+  | 'warning';
