+++ src/components/stable/gux-pagination/gux-pagination.types.ts
@@ -1,0 +1,8 @@
+export type GuxItemsPerPage = 25 | 50 | 75 | 100;
+
+export type GuxPaginationLayout = 'small' | 'expanded' | 'full';
+
+export type GuxPaginationState = {
+  currentPage: number;
+  itemsPerPage: number;
+};
