+++ src/components/stable/gux-simple-toast/gux-simple-toast.types.ts
@@ -1,0 +1,1 @@
+export type GuxSimpleToastAccent = 'neutral' | 'positive' | 'alert' | 'warning';
