+++ src/components/stable/gux-text-highlight/gux-text-highlight.types.ts
@@ -1,0 +1,1 @@
+export type GuxTextHighlightStrategy = 'start' | 'contains' | 'fuzzy';
